USB/IP: Warning fix. ​bugref:9861
svn:sync-xref-src-repo-rev: r142078

+++ src/VBox/Devices/USB/usbip/USBProxyDevice-usbip.cpp
@@ -1,4 +1,4 @@
-/* $Id: USBProxyDevice-usbip.cpp 87161 2021-01-04 12:45:18Z alexander.eichner@oracle.com $ */
+/* $Id: USBProxyDevice-usbip.cpp 87163 2021-01-04 13:04:49Z andreas.loeffler@oracle.com $ */
 /** @file
  * USB device proxy - USB/IP backend.
  */
@@ -997,7 +997,7 @@ static int usbProxyUsbIpRecvPdu(PUSBPROXYDEVUSBIP pProxyDevUsbIp, PUSBPROXYURBUS
                                     {
                                         if (RT_LIKELY(pProxyDevUsbIp->pUrbUsbIp->pVUsbUrb->cbData >= cbRet))
                                         {
-                                            pProxyDevUsbIp->pUrbUsbIp->pVUsbUrb->cbData = cbRet;
+                                            pProxyDevUsbIp->pUrbUsbIp->pVUsbUrb->cbData = (uint32_t)cbRet;
                                             usbProxyUsbIpRecvStateAdvance(pProxyDevUsbIp, USBPROXYUSBIPRECVSTATE_URB_BUFFER,
                                                                           pbData, pProxyDevUsbIp->BufRet.RetSubmit.u32ActualLength);
                                         }
@@ -1807,4 +1807,3 @@ extern const USBPROXYBACK g_USBProxyDeviceUsbIp =
     usbProxyUsbIpWakeup,
     0
 };
-
